+++ source/inputViewer.hpp
@@ -59,7 +59,7 @@ namespace baal
         virtual ~InputViewer(){}
 		
     	// ----- PUBLIC INPUTVIEWER METHODS -----
-        void handleData(float timestamp, projection* p, bool spiked, bool empty, Network* network, Neuron* postNeuron)
+        void handleData(double timestamp, projection* p, bool spiked, bool empty, Network* network, Neuron* postNeuron)
         {
         	if (!empty)
         	{
@@ -82,7 +82,7 @@ namespace baal
         }
 		
 		// ----- SETTERS -----
-        void setTimeWindow(float newWindow)
+        void setTimeWindow(double newWindow)
         {
             timeWindow = newWindow;
         }
@@ -122,7 +122,7 @@ namespace baal
                     axisX->setRange(maxX - timeWindow, maxX+1);
                     if (!points.isEmpty())
                     {
-                        auto firstToKeep = std::upper_bound(points.begin(), points.end(), points.back().x() - timeWindow, [](float timestamp, const QPointF& point) {
+                        auto firstToKeep = std::upper_bound(points.begin(), points.end(), points.back().x() - timeWindow, [](double timestamp, const QPointF& point) {
                             return timestamp < point.x();
                         });
                         points.remove(0, static_cast<int>(std::distance(points.begin(), firstToKeep)));
@@ -140,9 +140,9 @@ namespace baal
     	// ----- IMPLEMENTATION VARIABLES -----
         bool                  isClosed;
         bool                  openGL;
-        float                 timeWindow;
+        double                 timeWindow;
         QVector<QPointF>      points;
-        float                 maxX;
+        double                 maxX;
         int                   minY;
         int                   maxY;
         std::atomic_flag      atomicGuard;
